Combat Ready
Added: The Scout role now properly engages hostiles when in the same room as the rally flag (currently hard coded to 'Flag1')

Improved: Builder job allocation to dynamically scale the number of workers assigned to upgrading the room controller.

Improved:  Harvester drones will continue to mine if there is a container with available space on the same tile.